+++ controller/editorcontroller.php
@@ -13,7 +13,7 @@
  * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
  * For details, see the GNU AGPL at: http://www.gnu.org/licenses/agpl-3.0.html
  *
- * You can contact Ascensio System SIA at 17-2 Elijas street, Riga, Latvia, EU, LV-1021.
+ * You can contact Ascensio System SIA at 20A-12 Ernesta Birznieka-Upisha street, Riga, Latvia, EU, LV-1050.
  *
  * The interactive user interfaces in modified source and object code versions of the Program
  * must display Appropriate Legal Notices, as required under Section 5 of the GNU AGPL version 3.
@@ -185,28 +185,28 @@ class EditorController extends Controller {
      *
      * @param string $name - file name
      * @param string $dir - folder path
-     * @param string $token - access token
+     * @param string $shareToken - access token
      *
      * @return array
      *
      * @NoAdminRequired
      * @PublicPage
      */
-    public function create($name, $dir, $token = NULL) {
-        $this->logger->debug("Create: " . $name, array("app" => $this->appName));
+    public function create($name, $dir, $shareToken = NULL) {
+        $this->logger->debug("Create: $name", array("app" => $this->appName));
 
-        if (empty($token) && !$this->config->isUserAllowedToUse()) {
+        if (empty($shareToken) && !$this->config->isUserAllowedToUse()) {
             return ["error" => $this->trans->t("Not permitted")];
         }
 
-        if (empty($token)) {
+        if (empty($shareToken)) {
             $userId = $this->userSession->getUser()->getUID();
             $userFolder = $this->root->getUserFolder($userId);
         } else {
-            list ($userFolder, $error, $share) = $this->getNodeByToken($token);
+            list ($userFolder, $error, $share) = $this->getNodeByToken($shareToken);
 
             if (isset($error)) {
-                $this->logger->error("Create: " . $error, array("app" => $this->appName));
+                $this->logger->error("Create: $error", array("app" => $this->appName));
                 return ["error" => $error];
             }
 
@@ -214,8 +214,8 @@ class EditorController extends Controller {
                 return ["error" => $this->trans->t("You don't have enough permission to create")];
             }
 
-            if (!empty($token) && ($share->getPermissions() & Constants::PERMISSION_CREATE) === 0) {
-                $this->logger->error("Create in public folder without access: " . $fileId, array("app" => $this->appName));
+            if (!empty($shareToken) && ($share->getPermissions() & Constants::PERMISSION_CREATE) === 0) {
+                $this->logger->error("Create in public folder without access: $fileId", array("app" => $this->appName));
                 return ["error" => $this->trans->t("You do not have enough permissions to view the file")];
             }
         }
@@ -223,11 +223,11 @@ class EditorController extends Controller {
         $folder = $userFolder->get($dir);
 
         if ($folder === NULL) {
-            $this->logger->error("Folder for file creation was not found: " . $dir, array("app" => $this->appName));
+            $this->logger->error("Folder for file creation was not found: $dir", array("app" => $this->appName));
             return ["error" => $this->trans->t("The required folder was not found")];
         }
         if (!$folder->isCreatable()) {
-            $this->logger->error("Folder for file creation without permission: " . $dir, array("app" => $this->appName));
+            $this->logger->error("Folder for file creation without permission: $dir", array("app" => $this->appName));
             return ["error" => $this->trans->t("You don't have enough permission to create")];
         }
 
@@ -243,7 +243,7 @@ class EditorController extends Controller {
 
         $template = file_get_contents($templatePath);
         if (!$template) {
-            $this->logger->error("Template for file creation not found: " . $templatePath, array("app" => $this->appName));
+            $this->logger->error("Template for file creation not found: $templatePath", array("app" => $this->appName));
             return ["error" => $this->trans->t("Template not found")];
         }
 
@@ -254,7 +254,7 @@ class EditorController extends Controller {
 
             $file->putContent($template);
         } catch (NotPermittedException $e) {
-            $this->logger->error("Can't create file: " . $name, array("app" => $this->appName));
+            $this->logger->error("Can't create file: $name", array("app" => $this->appName));
             return ["error" => $this->trans->t("Can't create file")];
         }
 
@@ -280,17 +280,17 @@ class EditorController extends Controller {
      * Conversion file to Office Open XML format
      *
      * @param integer $fileId - file identifier
-     * @param string $token - access token
+     * @param string $shareToken - access token
      *
      * @return array
      *
      * @NoAdminRequired
      * @PublicPage
      */
-    public function convert($fileId, $token = NULL) {
-        $this->logger->debug("Convert: " . $fileId, array("app" => $this->appName));
+    public function convert($fileId, $shareToken = NULL) {
+        $this->logger->debug("Convert: $fileId", array("app" => $this->appName));
 
-        if (empty($token) && !$this->config->isUserAllowedToUse()) {
+        if (empty($shareToken) && !$this->config->isUserAllowedToUse()) {
             return ["error" => $this->trans->t("Not permitted")];
         }
 
@@ -300,15 +300,15 @@ class EditorController extends Controller {
             $userId = $user->getUID();
         }
 
-        list ($file, $error, $share) = empty($token) ? $this->getFile($userId, $fileId) : $this->getFileByToken($fileId, $token);
+        list ($file, $error, $share) = empty($shareToken) ? $this->getFile($userId, $fileId) : $this->getFileByToken($fileId, $shareToken);
 
         if (isset($error)) {
-            $this->logger->error("Convertion: " . $fileId . " " . $error, array("app" => $this->appName));
+            $this->logger->error("Convertion: $fileId $error", array("app" => $this->appName));
             return ["error" => $error];
         }
 
-        if (!empty($token) && ($share->getPermissions() & Constants::PERMISSION_CREATE) === 0) {
-            $this->logger->error("Convertion in public folder without access: " . $fileId, array("app" => $this->appName));
+        if (!empty($shareToken) && ($share->getPermissions() & Constants::PERMISSION_CREATE) === 0) {
+            $this->logger->error("Convertion in public folder without access: $fileId", array("app" => $this->appName));
             return ["error" => $this->trans->t("You do not have enough permissions to view the file")];
         }
 
@@ -316,12 +316,12 @@ class EditorController extends Controller {
         $ext = strtolower(pathinfo($fileName, PATHINFO_EXTENSION));
         $format = $this->config->FormatsSetting()[$ext];
         if (!isset($format)) {
-            $this->logger->info("Format for convertion not supported: " . $fileName, array("app" => $this->appName));
+            $this->logger->info("Format for convertion not supported: $fileName", array("app" => $this->appName));
             return ["error" => $this->trans->t("Format is not supported")];
         }
 
         if (!isset($format["conv"]) || $format["conv"] !== true) {
-            $this->logger->info("Conversion is not required: " . $fileName, array("app" => $this->appName));
+            $this->logger->info("Conversion is not required: $fileName", array("app" => $this->appName));
             return ["error" => $this->trans->t("Conversion is not required")];
         }
 
@@ -338,12 +338,11 @@ class EditorController extends Controller {
         $newFileUri;
         $documentService = new DocumentService($this->trans, $this->config);
         $key = $this->getKey($file);
-        $fileId = $file->getId();
-        $fileUrl = $this->getUrl($fileId, $token);
+        $fileUrl = $this->getUrl($file, $shareToken);
         try {
             $newFileUri = $documentService->GetConvertedUri($fileUrl, $ext, $internalExtension, $key);
         } catch (\Exception $e) {
-            $this->logger->error("GetConvertedUri: " . $fileId . " " . $e->getMessage(), array("app" => $this->appName));
+            $this->logger->error("GetConvertedUri: " . $file->getId() . " " . $e->getMessage(), array("app" => $this->appName));
             return ["error" => $e->getMessage()];
         }
 
@@ -355,7 +354,7 @@ class EditorController extends Controller {
         try {
             $newData = $documentService->Request($newFileUri);
         } catch (\Exception $e) {
-            $this->logger->error("Failed to download converted file: " . $newFileUri . " " . $e->getMessage(), array("app" => $this->appName));
+            $this->logger->error("Failed to download converted file: $newFileUri" . " " . $e->getMessage(), array("app" => $this->appName));
             return ["error" => $this->trans->t("Failed to download converted file")];
         }
 
@@ -367,7 +366,7 @@ class EditorController extends Controller {
 
             $file->putContent($newData);
         } catch (NotPermittedException $e) {
-            $this->logger->error("Can't create file: " . $newFileName, array("app" => $this->appName));
+            $this->logger->error("Can't create file: $newFileName", array("app" => $this->appName));
             return ["error" => $this->trans->t("Can't create file")];
         }
 
@@ -378,10 +377,116 @@ class EditorController extends Controller {
     }
 
     /**
+     * Save file to folder
+     *
+     * @param string $name - file name
+     * @param string $dir - folder path
+     * @param string $url - file url
+     *
+     * @return array
+     *
+     * @NoAdminRequired
+     */
+    public function save($name, $dir, $url) {
+        $this->logger->debug("Save: $name", array("app" => $this->appName));
+
+        if (!$this->config->isUserAllowedToUse()) {
+            return ["error" => $this->trans->t("Not permitted")];
+        }
+
+        $userId = $this->userSession->getUser()->getUID();
+        $userFolder = $this->root->getUserFolder($userId);
+
+        $folder = $userFolder->get($dir);
+
+        if ($folder === NULL) {
+            $this->logger->error("Folder for saving file was not found: $dir", array("app" => $this->appName));
+            return ["error" => $this->trans->t("The required folder was not found")];
+        }
+        if (!$folder->isCreatable()) {
+            $this->logger->error("Folder for saving file without permission: $dir", array("app" => $this->appName));
+            return ["error" => $this->trans->t("You don't have enough permission to create")];
+        }
+
+        $url = $this->config->ReplaceDocumentServerUrlToInternal($url);
+
+        try {
+            $documentService = new DocumentService($this->trans, $this->config);
+            $newData = $documentService->Request($url);
+        } catch (\Exception $e) {
+            $this->logger->error("Failed to download file for saving: $url " . $e->getMessage(), array("app" => $this->appName));
+            return ["error" => $this->trans->t("Download failed")];
+        }
+
+        $name = $folder->getNonExistingName($name);
+
+        try {
+            $file = $folder->newFile($name);
+
+            $file->putContent($newData);
+        } catch (NotPermittedException $e) {
+            $this->logger->error("Can't save file: $name", array("app" => $this->appName));
+            return ["error" => $this->trans->t("Can't create file")];
+        }
+
+        $fileInfo = $file->getFileInfo();
+
+        $result = Helper::formatFileInfo($fileInfo);
+        return $result;
+    }
+
+    /**
+     * Get presigned url to file
+     *
+     * @param string $filePath - file path
+     *
+     * @return array
+     *
+     * @NoAdminRequired
+     */
+    public function url($filePath) {
+        $this->logger->debug("Save: $name", array("app" => $this->appName));
+
+        if (!$this->config->isUserAllowedToUse()) {
+            return ["error" => $this->trans->t("Not permitted")];
+        }
+
+        $userId = $this->userSession->getUser()->getUID();
+        $userFolder = $this->root->getUserFolder($userId);
+
+        $file = $userFolder->get($filePath);
+
+        if ($file === NULL) {
+            $this->logger->error("File for generate presigned url was not found: $dir", array("app" => $this->appName));
+            return ["error" => $this->trans->t("File not found")];
+        }
+        if (!$file->isReadable()) {
+            $this->logger->error("Folder for saving file without permission: $dir", array("app" => $this->appName));
+            return ["error" => $this->trans->t("You do not have enough permissions to view the file")];
+        }
+
+        $fileName = $file->getName();
+        $ext = strtolower(pathinfo($fileName, PATHINFO_EXTENSION));
+        $fileUrl = $this->getUrl($file);
+
+        $result = [
+            "fileType" => $ext,
+            "url" => $fileUrl
+        ];
+
+        if (!empty($this->config->GetDocumentServerSecret())) {
+            $token = \Firebase\JWT\JWT::encode($result, $this->config->GetDocumentServerSecret());
+            $result["token"] = $token;
+        }
+
+        return $result;
+    }
+
+    /**
      * Print editor section
      *
      * @param integer $fileId - file identifier
-     * @param string $token - access token
+     * @param string $shareToken - access token
      * @param string $filePath - file path
      *
      * @return TemplateResponse|RedirectResponse
@@ -389,17 +494,17 @@ class EditorController extends Controller {
      * @NoAdminRequired
      * @NoCSRFRequired
      */
-    public function index($fileId, $token = NULL, $filePath = NULL) {
-        $this->logger->debug("Open: " . $fileId . " " . $filePath, array("app" => $this->appName));
+    public function index($fileId, $shareToken = NULL, $filePath = NULL) {
+        $this->logger->debug("Open: $fileId $filePath", array("app" => $this->appName));
 
-        if (empty($token) && !$this->userSession->isLoggedIn()) {
+        if (empty($shareToken) && !$this->userSession->isLoggedIn()) {
             $redirectUrl = $this->urlGenerator->linkToRoute("core.login.showLoginForm", [
                 "redirect_url" => $this->request->getRequestUri()
             ]);
             return new RedirectResponse($redirectUrl);
         }
 
-        if (empty($token) && !$this->config->isUserAllowedToUse()) {
+        if (empty($shareToken) && !$this->config->isUserAllowedToUse()) {
             return $this->renderError($this->trans->t("Not permitted"));
         }
 
@@ -414,7 +519,7 @@ class EditorController extends Controller {
             "documentServerUrl" => $documentServerUrl,
             "fileId" => $fileId,
             "filePath" => $filePath,
-            "token" => $token
+            "shareToken" => $shareToken
         ];
 
         $response = new TemplateResponse($this->appName, "editor", $params);
@@ -437,7 +542,7 @@ class EditorController extends Controller {
      * Print public editor section
      *
      * @param integer $fileId - file identifier
-     * @param string $token - access token
+     * @param string $shareToken - access token
      *
      * @return TemplateResponse
      *
@@ -445,8 +550,8 @@ class EditorController extends Controller {
      * @NoCSRFRequired
      * @PublicPage
      */
-    public function PublicPage($fileId, $token) {
-        return $this->index($fileId, $token);
+    public function PublicPage($fileId, $shareToken) {
+        return $this->index($fileId, $shareToken);
     }
 
     /**
@@ -454,7 +559,7 @@ class EditorController extends Controller {
      *
      * @param integer $fileId - file identifier
      * @param string $filePath - file path
-     * @param string $token - access token
+     * @param string $shareToken - access token
      * @param bool $desktop - desktop label
      *
      * @return array
@@ -462,9 +567,9 @@ class EditorController extends Controller {
      * @NoAdminRequired
      * @PublicPage
      */
-    public function config($fileId, $filePath = NULL, $token = NULL, $desktop = false) {
+    public function config($fileId, $filePath = NULL, $shareToken = NULL, $desktop = false) {
 
-        if (empty($token) && !$this->config->isUserAllowedToUse()) {
+        if (empty($shareToken) && !$this->config->isUserAllowedToUse()) {
             return ["error" => $this->trans->t("Not permitted")];
         }
 
@@ -474,10 +579,10 @@ class EditorController extends Controller {
             $userId = $user->getUID();
         }
 
-        list ($file, $error, $share) = empty($token) ? $this->getFile($userId, $fileId, $filePath) : $this->getFileByToken($fileId, $token);
+        list ($file, $error, $share) = empty($shareToken) ? $this->getFile($userId, $fileId, $filePath) : $this->getFileByToken($fileId, $shareToken);
 
         if (isset($error)) {
-            $this->logger->error("Config: " . $fileId . " " . $error, array("app" => $this->appName));
+            $this->logger->error("Config: $fileId $error", array("app" => $this->appName));
             return ["error" => $error];
         }
 
@@ -485,12 +590,11 @@ class EditorController extends Controller {
         $ext = strtolower(pathinfo($fileName, PATHINFO_EXTENSION));
         $format = $this->config->FormatsSetting()[$ext];
         if (!isset($format)) {
-            $this->logger->info("Format is not supported for editing: " . $fileName, array("app" => $this->appName));
+            $this->logger->info("Format is not supported for editing: $fileName", array("app" => $this->appName));
             return ["error" => $this->trans->t("Format is not supported")];
         }
 
-        $fileId = $file->getId();
-        $fileUrl = $this->getUrl($fileId, $token);
+        $fileUrl = $this->getUrl($file, $shareToken);
         $key = $this->getKey($file);
 
         $params = [
@@ -503,11 +607,12 @@ class EditorController extends Controller {
             ],
             "documentType" => $format["type"],
             "editorConfig" => [
-                "lang" => str_replace("_", "-", \OC::$server->getL10NFactory("")->get("")->getLanguageCode())
+                "lang" => str_replace("_", "-", \OC::$server->getL10NFactory("")->get("")->getLanguageCode()),
+                "region" => str_replace("_", "-", \OC::$server->getL10NFactory("")->findLocale())
             ]
         ];
 
-        $permissions_modifyFilter = $this->config->getSystemValue($this->config->_permissions_modifyFilter);
+        $permissions_modifyFilter = $this->config->GetSystemValue($this->config->_permissions_modifyFilter);
         if (isset($permissions_modifyFilter)) {
             $params["document"]["permissions"]["modifyFilter"] = $permissions_modifyFilter;
         }
@@ -522,6 +627,9 @@ class EditorController extends Controller {
 			$editable = $event->getArgument('editable');
 		}
 
+        $params["document"]["permissions"]["edit"] = $editable;
+
+
         if ($editable && $canEdit) {
             $ownerId = NULL;
             $owner = $file->getOwner();
@@ -529,7 +637,7 @@ class EditorController extends Controller {
                 $ownerId = $owner->getUID();
             }
 
-            $hashCallback = $this->crypt->GetHash(["fileId" => $fileId, "ownerId" => $ownerId, "token" => $token, "action" => "track"]);
+            $hashCallback = $this->crypt->GetHash(["fileId" => $file->getId(), "ownerId" => $ownerId, "shareToken" => $shareToken, "action" => "track"]);
             $callback = $this->urlGenerator->linkToRouteAbsolute($this->appName . ".callback.track", ["doc" => $hashCallback]);
 
             if (!empty($this->config->GetStorageUrl())) {
@@ -564,12 +672,10 @@ class EditorController extends Controller {
 
         $folderLink = NULL;
 
-        if (!empty($token)) {
+        if (!empty($shareToken)) {
             if (method_exists($share, "getHideDownload") && $share->getHideDownload()) {
-                $params["document"]["permissions"] = [
-                    "download" => false,
-                    "print" => false
-                ];
+                $params["document"]["permissions"]["download"] = false;
+                $params["document"]["permissions"]["print"] = false;
             }
 
             $node = $share->getNode();
@@ -580,7 +686,7 @@ class EditorController extends Controller {
                     $linkAttr = [
                         "path" => $folderPath,
                         "scrollto" => $file->getName(),
-                        "token" => $token
+                        "token" => $shareToken
                     ];
                     $folderLink = $this->urlGenerator->linkToRouteAbsolute("files_sharing.sharecontroller.showShare", $linkAttr);
                 }
@@ -611,12 +717,18 @@ class EditorController extends Controller {
 
         $params = $this->setCustomization($params);
 
+        $params = $this->setWatermark($params, !empty($shareToken), $userId, $file);
+
+        if ($this->config->UseDemo()) {
+            $params["editorConfig"]["tenant"] = $this->config->GetSystemValue("instanceid", true);
+        }
+
         if (!empty($this->config->GetDocumentServerSecret())) {
             $token = \Firebase\JWT\JWT::encode($params, $this->config->GetDocumentServerSecret());
             $params["token"] = $token;
         }
 
-        $this->logger->debug("Config is generated for: " . $fileId . " with key " . $key, array("app" => $this->appName));
+        $this->logger->debug("Config is generated for: $fileId with key $key", array("app" => $this->appName));
 
         return $params;
     }
@@ -635,10 +747,15 @@ class EditorController extends Controller {
             return [NULL, $this->trans->t("FileId is empty"), NULL];
         }
 
-        $files = $this->root->getUserFolder($userId)->getById($fileId);
+        try {
+            $files = $this->root->getUserFolder($userId)->getById($fileId);
+        } catch (\Exception $e) {
+            $this->logger->error("getFile: $fileId " . $e->getMessage(), array("app" => $this->appName));
+            return [NULL, $this->trans->t("Invalid request"), NULL];
+        }
 
         if (empty($files)) {
-            $this->logger->info("Files not found: " . $fileId, array("app" => $this->appName));
+            $this->logger->info("Files not found: $fileId", array("app" => $this->appName));
             return [NULL, $this->trans->t("File not found"), NULL];
         }
 
@@ -664,22 +781,27 @@ class EditorController extends Controller {
      * Getting file by token
      *
      * @param integer $fileId - file identifier
-     * @param string $token - access token
+     * @param string $shareToken - access token
      *
      * @return array
      */
-    private function getFileByToken($fileId, $token) {
-        list ($node, $error, $share) = $this->getNodeByToken($token);
+    private function getFileByToken($fileId, $shareToken) {
+        list ($node, $error, $share) = $this->getNodeByToken($shareToken);
 
         if (isset($error)) {
             return [NULL, $error, NULL];
         }
 
         if ($node instanceof Folder) {
-            $files = $node->getById($fileId);
+            try {
+                $files = $node->getById($fileId);
+            } catch (\Exception $e) {
+                $this->logger->error("getFileByToken: $fileId " . $e->getMessage(), array("app" => $this->appName));
+                return [NULL, $this->trans->t("Invalid request"), NULL];
+            }
 
             if (empty($files)) {
-                $this->logger->info("Files not found: " . $fileId, array("app" => $this->appName));
+                $this->logger->info("Files not found: $fileId", array("app" => $this->appName));
                 return [NULL, $this->trans->t("File not found"), NULL];
             }
             $file = $files[0];
@@ -693,12 +815,12 @@ class EditorController extends Controller {
     /**
      * Getting file by token
      *
-     * @param string $token - access token
+     * @param string $shareToken - access token
      *
      * @return array
      */
-    private function getNodeByToken($token) {
-        list ($share, $error) = $this->getShare($token);
+    private function getNodeByToken($shareToken) {
+        list ($share, $error) = $this->getShare($shareToken);
 
         if (isset($error)) {
             return [NULL, $error, NULL];
@@ -721,18 +843,18 @@ class EditorController extends Controller {
     /**
      * Getting share by token
      *
-     * @param string $token - access token
+     * @param string $shareToken - access token
      *
      * @return array
      */
-    private function getShare($token) {
-        if (empty($token)) {
+    private function getShare($shareToken) {
+        if (empty($shareToken)) {
             return [NULL, $this->trans->t("FileId is empty")];
         }
 
         $share;
         try {
-            $share = $this->shareManager->getShareByToken($token);
+            $share = $this->shareManager->getShareByToken($shareToken);
         } catch (ShareNotFound $e) {
             $this->logger->error("getShare error: " . $e->getMessage(), array("app" => $this->appName));
             $share = NULL;
@@ -759,7 +881,7 @@ class EditorController extends Controller {
      * @return string
      */
     private function getKey($file) {
-        $instanceId = $this->config->getSystemValue("instanceid", true);
+        $instanceId = $this->config->GetSystemValue("instanceid", true);
 
         $fileId = $file->getId();
 
@@ -771,12 +893,12 @@ class EditorController extends Controller {
     /**
      * Generate secure link to download document
      *
-     * @param integer $fileId - file identifier
-     * @param string $token - access token
+     * @param integer $file - file
+     * @param string $shareToken - access token
      *
      * @return string
      */
-    private function getUrl($fileId, $token = NULL) {
+    private function getUrl($file, $shareToken = NULL) {
 
         $user = $this->userSession->getUser();
         $userId = NULL;
@@ -784,7 +906,13 @@ class EditorController extends Controller {
             $userId = $user->getUID();
         }
 
-        $hashUrl = $this->crypt->GetHash(["fileId" => $fileId, "userId" => $userId, "token" => $token, "action" => "download"]);
+        $ownerId = NULL;
+        $owner = $file->getOwner();
+        if (!empty($owner)) {
+            $ownerId = $owner->getUID();
+        }
+
+        $hashUrl = $this->crypt->GetHash(["fileId" => $file->getId(), "userId" => $userId, "ownerId" => $ownerId, "shareToken" => $shareToken, "action" => "download"]);
 
         $fileUrl = $this->urlGenerator->linkToRouteAbsolute($this->appName . ".callback.download", ["doc" => $hashUrl]);
 
@@ -831,32 +959,147 @@ class EditorController extends Controller {
 
         /* from system config */
 
-        $customer = $this->config->getSystemValue($this->config->_customization_customer);
+        $customer = $this->config->GetSystemValue($this->config->_customization_customer);
         if (isset($customer)) {
             $params["editorConfig"]["customization"]["customer"] = $customer;
         }
 
-        $feedback = $this->config->getSystemValue($this->config->_customization_feedback);
+        $feedback = $this->config->GetSystemValue($this->config->_customization_feedback);
         if (isset($feedback)) {
             $params["editorConfig"]["customization"]["feedback"] = $feedback;
         }
 
-        $loaderLogo = $this->config->getSystemValue($this->config->_customization_loaderLogo);
+        $loaderLogo = $this->config->GetSystemValue($this->config->_customization_loaderLogo);
         if (isset($loaderLogo)) {
             $params["editorConfig"]["customization"]["loaderLogo"] = $loaderLogo;
         }
 
-        $loaderName = $this->config->getSystemValue($this->config->_customization_loaderName);
+        $loaderName = $this->config->GetSystemValue($this->config->_customization_loaderName);
         if (isset($loaderName)) {
             $params["editorConfig"]["customization"]["loaderName"] = $loaderName;
         }
 
-        $logo = $this->config->getSystemValue($this->config->_customization_logo);
+        $logo = $this->config->GetSystemValue($this->config->_customization_logo);
         if (isset($logo)) {
             $params["editorConfig"]["customization"]["logo"] = $logo;
         }
 
         return $params;
+    }
+
+    /**
+     * Set watermark parameters
+     *
+     * @param array params - file parameters
+     * @param bool isPublic - with access token
+     * @param string userId - user identifier
+     * @param string file - file
+     *
+     * @return array
+     */
+    private function setWatermark($params, $isPublic, $userId, $file) {
+        $watermarkTemplate = $this->getWatermarkText($isPublic, $userId, $file,
+            $params["document"]["permissions"]["edit"] !== false,
+            !array_key_exists("download", $params["document"]["permissions"]) || $params["document"]["permissions"]["download"] !== false);
+
+        if ($watermarkTemplate !== false) {
+            $replacements = [
+                "userId" => $userId,
+                "date" => (new \DateTime())->format("Y-m-d H:i:s"),
+                "themingName" => \OC::$server->getThemingDefaults()->getName()
+            ];
+            $watermarkTemplate = preg_replace_callback("/{(.+?)}/", function($matches) use ($replacements)
+                {
+                    return $replacements[$matches[1]];
+                }, $watermarkTemplate);
+
+            $params["document"]["options"] = [
+                "watermark_on_draw" => [
+                    "align" => 1,
+                    "height" => 100,
+                    "paragraphs" => array([
+                        "align" => 2,
+                        "runs" => array([
+                            "fill" => [182, 182, 182],
+                            "font-size" => 70,
+                            "text" => $watermarkTemplate,
+                        ])
+                    ]),
+                    "rotate" => -45,
+                    "width" => 250,
+                ]
+            ];
+        }
+
+        return $params;
+    }
+
+    /**
+     * Should watermark
+     *
+     * @param bool isPublic - with access token
+     * @param string userId - user identifier
+     * @param string file - file
+     * @param bool canEdit - edit permission
+     * @param bool canDownload - download permission
+     *
+     * @return bool|string
+     */
+    private function getWatermarkText($isPublic, $userId, $file, $canEdit, $canDownload) {
+        $watermarkSettings = $this->config->GetWatermarkSettings();
+        if (!$watermarkSettings["enabled"]) {
+            return false;
+        }
+
+        $watermarkText = $watermarkSettings["text"];
+        $fileId = $file->getId();
+
+        if ($isPublic) {
+            if ($watermarkSettings["linkAll"]) {
+                return $watermarkText;
+            }
+            if ($watermarkSettings["linkRead"] && !$canEdit) {
+                return $watermarkText;
+            }
+            if ($watermarkSettings["linkSecure"] && !$canDownload) {
+                return $watermarkText;
+            }
+            if ($watermarkSettings["linkTags"]) {
+                $tags = $watermarkSettings["linkTagsList"];
+                $fileTags = \OC::$server->getSystemTagObjectMapper()->getTagIdsForObjects([$fileId], "files")[$fileId];
+                foreach ($fileTags as $tagId) {
+                    if (in_array($tagId, $tags, true)) {
+                        return $watermarkText;
+                    }
+                }
+            }
+        } else {
+            if ($watermarkSettings["shareAll"] && $file->getOwner()->getUID() !== $userId) {
+                return $watermarkText;
+            }
+            if ($watermarkSettings["shareRead"] && !$canEdit) {
+                return $watermarkText;
+            }
+        }
+        if ($watermarkSettings["allGroups"]) {
+            $groups = $watermarkSettings["allGroupsList"];
+            foreach ($groups as $group) {
+                if (\OC::$server->getGroupManager()->isInGroup($userId, $group)) {
+                    return $watermarkText;
+                }
+            }
+        }
+        if ($watermarkSettings["allTags"]) {
+            $tags = $watermarkSettings["allTagsList"];
+            $fileTags = \OC::$server->getSystemTagObjectMapper()->getTagIdsForObjects([$fileId], "files")[$fileId];
+            foreach ($fileTags as $tagId) {
+                if (in_array($tagId, $tags, true)) {
+                    return $watermarkText;
+                }
+            }
+        }
+
+        return false;
     }
 
     /**