+++ js/main.js
@@ -12,7 +12,7 @@
  * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
  * For details, see the GNU AGPL at: http://www.gnu.org/licenses/agpl-3.0.html
  *
- * You can contact Ascensio System SIA at 17-2 Elijas street, Riga, Latvia, EU, LV-1021.
+ * You can contact Ascensio System SIA at 20A-12 Ernesta Birznieka-Upisha street, Riga, Latvia, EU, LV-1050.
  *
  * The interactive user interfaces in modified source and object code versions of the Program
  * must display Appropriate Legal Notices, as required under Section 5 of the GNU AGPL version 3.
@@ -51,7 +51,7 @@
         };
 
         if ($("#isPublic").val()) {
-            createData.token = encodeURIComponent($("#sharingToken").val());
+            createData.shareToken = encodeURIComponent($("#sharingToken").val());
         }
 
         $.post(OC.generateUrl("apps/" + OCA.Onlyoffice.AppName + "/ajax/new"),
@@ -87,9 +87,9 @@
             });
 
         if ($("#isPublic").val()) {
-            url = OC.generateUrl("apps/" + OCA.Onlyoffice.AppName + "/s/{token}?fileId={fileId}",
+            url = OC.generateUrl("apps/" + OCA.Onlyoffice.AppName + "/s/{shareToken}?fileId={fileId}",
                 {
-                    token: encodeURIComponent($("#sharingToken").val()),
+                    shareToken: encodeURIComponent($("#sharingToken").val()),
                     fileId: fileId
                 });
         }
@@ -117,7 +117,7 @@
         };
 
         if ($("#isPublic").val()) {
-            convertData.token = encodeURIComponent($("#sharingToken").val());
+            convertData.shareToken = encodeURIComponent($("#sharingToken").val());
         }
 
         $.post(OC.generateUrl("apps/" + OCA.Onlyoffice.AppName + "/ajax/convert"),