+++ l10n/fr.js
@@ -53,6 +53,11 @@ OC.L10N.register(
     "Display the header more compact" : "Afficher l'en-tête plus compact",
     "Display Feedback & Support menu button" : "Afficher le bouton du menu 'Feedback & Support'",
     "Display Help menu button" : "Afficher le bouton du menu 'Aide'",
-    "Display toolbar tabs" : "Afficher les onglets de la barre d'outils"
+    "Display toolbar tabs" : "Afficher les onglets de la barre d'outils",
+    "Save as" : "Enregistrer sous",
+    "File saved" : "Le fichier a été enregistré",
+    "Insert image" : "Insérer une image",
+    "Select recipients" : "Sélectionner les destinataires",
+    "Select tag" : "Sélectionner une étiquette"
 },
 "nplurals=2; plural=(n != 1);");